Integrate stop callback and RGB565 frames; drop diagnostics
Brings in the two audio/compat commits from the port branch and removes the
temporary in-game-silence diagnostics.

- media.rs / backend audio.rs: adopt the interrupted-stop callback. MC_mdaStop
  now fires the clip callback with event=-1 when a playback was actually
  running, tracked by a play_generation / playing pair that also fences stale
  completion callbacks. This lets a title's sound state machine tear the
  current BGM down and start the next one, which is what the looping title BGM
  never reached in-game.
- Android frames go out as RGB565 short[] rather than ARGB_8888 int[]: half the
  data and a direct match for an RGB_565 Bitmap. The RGB565 change is applied to
  MainActivity by hand so the keypad-interruption and audio pause/resume already
  on this branch are kept.
- Remove the AUDIO_DIAG markers and the sound state-machine PC probe now that
  the cause is understood.

+++ android/app/src/main/java/com/jjongjjongs/wiemobile/NativeBridge.java
@@ -51,9 +51,9 @@ final class NativeBridge {
 
     /**
      * @return {@code null} when nothing new was painted, otherwise
-     *         {@code {width, height, ARGB_8888 pixels...}}
+     *         {@code {width, height, RGB565 pixels...}}
      */
-    static native int[] nativeFrame();
+    static native short[] nativeFrame();
 
     /**
      * @return the next queued audio or vibration command, or {@code null}
